chore: split out ChainDropdown from ChainDropdownBase
