+++ src/components/ChainDropdown/ChainDropdown.tsx
@@ -1,40 +1,13 @@
-import { ChevronDownIcon } from '@chakra-ui/icons'
-import type { ButtonProps, MenuProps } from '@chakra-ui/react'
-import {
-  Button,
-  Flex,
-  Menu,
-  MenuButton,
-  MenuItemOption,
-  MenuList,
-  MenuOptionGroup,
-} from '@chakra-ui/react'
+import { MenuItemOption } from '@chakra-ui/react'
 import type { ChainId } from '@shapeshiftoss/caip'
-import { useCallback, useMemo } from 'react'
-import { useTranslate } from 'react-polyglot'
-import { Amount } from 'components/Amount/Amount'
-import { IconCircle } from 'components/IconCircle'
-import { GridIcon } from 'components/Icons/GridIcon'
-import {
-  selectPortfolioTotalBalanceByChainIdIncludeStaking,
-  selectPortfolioTotalUserCurrencyBalanceExcludeEarnDupes,
-} from 'state/slices/selectors'
+import { useMemo } from 'react'
+import { selectPortfolioTotalBalanceByChainIdIncludeStaking } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
 
+import { ChainDropdownBase, type ChainDropdownBaseProps } from './ChainDropdownBase'
 import { ChainRow } from './ChainRow'
 
-type ChainDropdownProps = {
-  chainId?: ChainId
-  onClick: (arg: ChainId | undefined) => void
-  chainIds: ChainId[]
-  showAll?: boolean
-  includeBalance?: boolean
-  buttonProps?: ButtonProps
-} & Omit<MenuProps, 'children'>
-
-const width = { base: 'full', md: 'auto' }
-
-const chevronDownIcon = <ChevronDownIcon />
+type ChainDropdownProps = Omit<ChainDropdownBaseProps, 'chainOptions'>
 
 const ChainRowWithBalance = ({ chainId }: { chainId: ChainId }) => {
   const filter = useMemo(() => ({ chainId }), [chainId])
@@ -45,21 +18,8 @@ const ChainRowWithBalance = ({ chainId }: { chainId: ChainId }) => {
   return <ChainRow chainId={chainId} fiatBalance={chainFiatBalance} />
 }
 
-export const ChainDropdown: React.FC<ChainDropdownProps> = ({
-  chainIds,
-  chainId,
-  onClick,
-  showAll,
-  includeBalance,
-  buttonProps,
-  ...menuProps
-}) => {
-  const totalPortfolioUserCurrencyBalance = useAppSelector(
-    selectPortfolioTotalUserCurrencyBalanceExcludeEarnDupes,
-  )
-  const translate = useTranslate()
-
-  const renderChains = useMemo(() => {
+export const ChainDropdown: React.FC<ChainDropdownProps> = ({ chainIds, ...props }) => {
+  const chainOptions = useMemo(() => {
     return chainIds.map(chainId => (
       <MenuItemOption value={chainId} key={chainId}>
         <ChainRowWithBalance chainId={chainId} />
@@ -67,29 +27,5 @@ export const ChainDropdown: React.FC<ChainDropdownProps> = ({
     ))
   }, [chainIds])
 
-  const onChange = useCallback((value: string | string[]) => onClick(value as ChainId), [onClick])
-
-  return (
-    <Menu {...menuProps}>
-      <MenuButton width={width} as={Button} rightIcon={chevronDownIcon} {...buttonProps}>
-        {chainId ? <ChainRow chainId={chainId} /> : translate('common.allChains')}
-      </MenuButton>
-      <MenuList zIndex='banner'>
-        <MenuOptionGroup type='radio' value={chainId} onChange={onChange}>
-          {showAll && (
-            <MenuItemOption value=''>
-              <Flex alignItems='center' gap={4}>
-                <IconCircle boxSize='24px'>
-                  <GridIcon />
-                </IconCircle>
-                {translate('common.allChains')}
-                <Amount.Fiat ml='auto' value={totalPortfolioUserCurrencyBalance} />
-              </Flex>
-            </MenuItemOption>
-          )}
-          {renderChains}
-        </MenuOptionGroup>
-      </MenuList>
-    </Menu>
-  )
+  return <ChainDropdownBase chainOptions={chainOptions} chainIds={chainIds} {...props} />
 }
